Top Navbar & Bottom Navbar components added
